Struct dzahui::logger::LogWriter

pub struct LogWriter {
    pub log_path: PathBuf,
    pub line_count: u64,
    pub line_maximum: u64,
    pub current_log_file_number: i64,
    pub log_file: File,
    pub rx: Receiver<String>,
}

General Information

Struct that writes logs to files. Changes files when maximum number of lines has been reached.

Fields

log_path - Direction in which files will be generated

line_count - Number of lines currently written

line_maximum - Maximum number of lines per file

current_log_file_number - Number of current file in which logs are being written

log_file - Current log file

rx - Used for internal communication between log, warn and error in this structure

Fields
log_path: PathBufline_count: u64line_maximum: u64current_log_file_number: i64log_file: Filerx: Receiver<String>
Implementations

impl LogWriter

pub fn new(
log_path: PathBuf,
rx: Receiver<String>,
line_maximum: u64
) -> LogWriter

General Information

Creates new instance of LogWriter while taking into account previous log files, last log file and state of last log file

Parameters

log_path - A path-like object where all logs remain

rx - A receiver for internal communication

line_maximum - Maximum number of lines for log file

pub fn run(&mut self)

General Information

Writes to log file and changes internal values like line number and log file if necessary

Parameters

&mut self - A mutable reference to write and change internal state