Enum dzahui::solvers::fem::Solver

pub enum Solver {
    DiffussionSolverTimeIndependent(DiffussionParamsTimeIndependent),
    DiffussionSolverTimeDependent(DiffussionParamsTimeDependent),
    Stokes1DSolver(StokesParams1D),
    Stokes2DSolver(StokesParams2D),
    None,
}

General Information

An enum representing every equation implemented by this library. Used as a way of representing an equation without having to create an instance of a solver of it. May be deprecated in the future in favor of a intercative approach from window interface.

Arms

DiffussionSolverTimeIndependent - Diffusion equation solver representation.

DiffussionSolverTimeDependent - Diffusion equation solver with time derivative representation.

Stokes1DSolver - Stokes in 1D solver representation.

Stokes2DSolver - Stokes in 2D solver representation.

None - Visuallization of mesh in simulation. No equation attached.

Variants

DiffussionSolverTimeIndependent(DiffussionParamsTimeIndependent)

DiffussionSolverTimeDependent(DiffussionParamsTimeDependent)

Stokes1DSolver(StokesParams1D)

Stokes2DSolver(StokesParams2D)

None